Drop --insecure/SECURE option + instead depend on debian-archive-keyring
When deploying Debian/squeeze from Debian/wheezy it fails with:

| Release signed by unknown key (key id AED4B06F473041FA)

and as Ubuntu provides grml-debootstrap in universe but its
/etc/apt/trusted.gpg doesn't include Debian's repos keys it also
fails there.  As soon as
/usr/share/keyrings/debian-archive-keyring.gpg is available we
don't have to take care of it, so just drop the additional code
and deprecate the --insecure cmdline option and SECURE config
file option and instead depend on the debian-archive-keyring
package.