don't fail to ask for a password if there is no TTY
When running grml-debootrap over ssh like
  ssh root@grml grml-debootstrap ...
the password promt for the root user would fail, as you cannot stty
without a tty. The user should actually have called
  ssh -t root@grml grml-debootstrap ...
which would have allocated a TTY and all would have worked.

Display a warning to the user and fall back to displaying the entered
password when he continues to use grml-debootstrap w/o a TTY.

Closes: issue1200