Remove the hardcoded apt pinning from --backportrepos and add a release check for it
Debian backports repository provides a default apt pinning value
from 100 and all changes must be done by the user to lift up this.
Instructions to change this, can be found on the website
(http://backports.debian.org).
But the side effect is by enabling the feature with release version
squeeze, that it installs the kernel headers from backports too,
because by the internal kernel and kernelheader routines
(image-2.6*.).