packer: improve support for testing local grml-debootstrap version
Sadly packer lacks an option to copy an entire directory
*excluding* certain files, otherwise we could just copy the whole
grml-debootstrap.git directory. But as to the packer_cache
directory is present in grml-debootstrap.git directory during
packer execution this makes no sense. So use a workaround to get
the same behaviour. Allow testing of grml-debootstrap without
having to commit and push stuff to git first, so one can hack on
stuff, test and only then commit and push it.