Docker setup for automated builds + environment for testing
It would be nice to automatically execute a wide range of tests
on each commit to the grml-debootstrap repository. We aren't
there yet, but a combination of docker and packer builds should
provide a decent starting point.

This change includes a Dockerfile to provide a public available
grml-debootstrap docker environment via the Docker Hub Registry
as well as scripts and tests for further exploration.

Please refer to the included README or visit
https://registry.hub.docker.com/u/mika/grml-debootstrap/ for
further information.