Prepare switch from legacy network interfaces to systemd-networkd
By default grml-debootstrap used the host /etc/network/interfaces file
to configure the network of the target system. This might have been ok,
when grml-debootrap was primarily run form a Grml live system. But this
might expose to much unwanted information about the host system when run
without it.

If the user really needs to copy any files from the host to the target
system, she should use a post-script.

While at it we also removed copying the interfaces.examples
functionality.

If no networking option is given, the --defaultinterfaces option will be
the new default.

We now provide a systemd-networkd configuration enabling DHCP on all
ethernet network devices next to the legacy /etc/network/interfaces
which only enables DHCP on eth0 (which does not work when using the
predictable devices names).

NOTE: systemd-networkd is currently not enabled by default.