Use bullseye as default Debian release
We want to use Debian/bullseye as default release within the
grml-debootstrap version that's going to be shipped with bullseye,
so let's be prepared for this.

FTR: *not* modifying the Travis CI related configuration files with this
change, as the unstable builds on Travis CI are known to be unreliable.
Travis CI also seems to be a dead end for us, so it's not worth putting
any further efforts into it, at least for the time being.