Implement option vmtype to select between raw or qcow2
qcow2 is a growable format not taking so much space from the beginning.
Instead of creating a raw image and converting it to qcow2 with
`qemu-img convert`

    $ qemu-img convert -f raw -O qcow2 debian.img debian.qcow2

add a switch `--vmtype` allowing to specify the type of the virtual
machine file, which is used in `qemu-img create -f`.

It looks like though, installing the system into that file using NBD is
twice as slow.

    $ sudo ~/src/grml-debootstrap/grml-debootstrap --arch amd64 --release sid --filesystem ext4 --vmfile