tests: Try to capture screenshot via VNC if serial-console-connection fails
While debugging the failing GitHub Actions (see #278), we noticed that
it might serve useful to have a screenshot of the booted system
available. We can capture such a screenshot via vncsnapshot.

NOTE: we move the files inside `results` only if the directory as such
exists in $PWD, otherwise we'd fail the build when being executed as
`tests/build-vm-and-test.sh test` from inside the grml-debootstrap.git.
This happens when manually reproducing the GH tests, though then
`build-vm-and-test.sh` is supposed to be executed from the parent
directory of the `tests` directory (being the grml-debootstrap.git
checkout).